Fix issue with Ember 2.5.0-beta.1.
Ember 2.5.0-beta.1 includes a "local lookup" feature, that calls
`componentFor` twice with different arguments:

1  `name, owner, { source }` -- Added options with `source` so that
  resolver can do local lookup logic based on the invoking template
  location.
2. `name, owner, { source: undefined }` -- Source is undefined meaning
  that we should look for the "global" component.

The `componentFor` method is paired with `layoutFor` to return the
componet/layout pair. When the normal implementation of `layoutFor`
is called in scenario 1 above, it receives a `source` param, but the
default ember-cli resolver is not local lookup aware so `layoutFor`
with a `source` option always returns undefined. Unfortunately, our
override was not properly passing along all the arguments that we were
invoked with, so the lookup for item 1) above was returning the
component (since the `source` param was not found in options it was
assumed to be the **global** lookup).

This meant that we would return the Component class from the "local
lookup" version of `componentFor`, but that the `layoutFor` invocation
would **never** find a template.

**tldr;** Always call `this._super` with all the arguments...