Remove the dependency on jQuery
This commit removes the dependency on jQuery and updates the
class name of the Ember root element when routes are activated
and/or deactivated.

It also does not automatically uses the `body` tag by default,
but "intelligently" tries to find the `appRoot` of the Ember
application by querying the container of the application for
the root element.

Additionally, since the operations on the document are purely
by setting and removing strings on the node attributes, this
should support FastBoot as well.